Move more inner classes to internal packages.
