+++ jpa/src/main/java/org/jdbi/v3/jpa/BindJpa.java
@@ -13,66 +13,17 @@
  */
 package org.jdbi.v3.jpa;
 
-import java.lang.annotation.Annotation;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
-import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
-import java.lang.reflect.Parameter;
 
-import java.lang.reflect.Type;
-import org.jdbi.v3.sqlobject.customizer.SqlStatementCustomizerFactory;
+import org.jdbi.v3.jpa.internal.BindJpaFactory;
 import org.jdbi.v3.sqlobject.customizer.SqlStatementCustomizingAnnotation;
-import org.jdbi.v3.sqlobject.customizer.SqlStatementParameterCustomizer;
 
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ ElementType.PARAMETER })
-@SqlStatementCustomizingAnnotation(BindJpa.Factory.class)
+@SqlStatementCustomizingAnnotation(BindJpaFactory.class)
 public @interface BindJpa {
     String value() default "";
-
-    class Factory implements SqlStatementCustomizerFactory {
-        @Override
-        public SqlStatementParameterCustomizer createForParameter(Annotation annotation,
-                                                                  Class<?> sqlObjectType,
-                                                                  Method method,
-                                                                  Parameter param,
-                                                                  int index,
-                                                                  Type type) {
-            BindJpa bind = (BindJpa) annotation;
-            final String prefix;
-            if (bind.value().isEmpty()) {
-                prefix = "";
-            } else {
-                prefix = bind.value() + ".";
-            }
-            return (stmt, arg) -> {
-                JpaClass<?> jpaClass = JpaClass.get(arg.getClass());
-                for (JpaMember member : jpaClass.members()) {
-                    stmt.bindByType(
-                            prefix + member.getColumnName(),
-                            readMember(arg, member),
-                            member.getType());
-                }
-            };
-        }
-
-        private static Object readMember(Object entity, JpaMember member) {
-            try {
-                return member.read(entity);
-            } catch (IllegalAccessException e) {
-                String message = String.format(
-                        "Unable to access property value for column %s",
-                        member.getColumnName());
-                throw new EntityMemberAccessException(message, e);
-            } catch (InvocationTargetException e) {
-                String message = String.format(
-                        "Exception thrown in accessor method for column %s",
-                        member.getColumnName());
-                throw new EntityMemberAccessException(message, e);
-            }
-        }
-    }
 }
