+++ sqlobject/src/main/java/org/jdbi/v3/sqlobject/customizer/Bind.java
@@ -13,23 +13,19 @@
  */
 package org.jdbi.v3.sqlobject.customizer;
 
-import java.lang.annotation.Annotation;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
-import java.lang.reflect.Method;
-import java.lang.reflect.Parameter;
-import java.lang.reflect.Type;
 
-import org.jdbi.v3.sqlobject.internal.ParameterUtil;
+import org.jdbi.v3.sqlobject.customizer.internal.BindFactory;
 
 /**
  * Binds the annotated argument as a named parameter, and as a positional parameter.
  */
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ElementType.PARAMETER})
-@SqlStatementCustomizingAnnotation(Bind.Factory.class)
+@SqlStatementCustomizingAnnotation(BindFactory.class)
 public @interface Bind
 {
     String NO_VALUE = "";
@@ -41,23 +37,4 @@ public @interface Bind
      * @return the name to which the argument will be bound.
      */
     String value() default NO_VALUE;
-
-    class Factory implements SqlStatementCustomizerFactory {
-        @Override
-        public SqlStatementParameterCustomizer createForParameter(Annotation annotation,
-                                                                  Class<?> sqlObjectType,
-                                                                  Method method,
-                                                                  Parameter param,
-                                                                  int index,
-                                                                  Type type) {
-            Bind b = (Bind) annotation;
-            String nameFromAnnotation = b == null ? NO_VALUE : b.value();
-            final String name = ParameterUtil.getParameterName(b, nameFromAnnotation, param);
-
-            return (stmt, arg) -> {
-                stmt.bindByType(index, arg, type);
-                stmt.bindByType(name, arg, type);
-            };
-        }
-    }
 }
