+++ sqlobject/src/main/java/org/jdbi/v3/sqlobject/customizer/internal/BindFactory.java
@@ -11,32 +11,32 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jdbi.v3.sqlobject.statement;
-
-import java.lang.reflect.Type;
-import org.jdbi.v3.sqlobject.customizer.Bind;
-import org.jdbi.v3.sqlobject.customizer.SqlStatementCustomizerFactory;
-import org.jdbi.v3.sqlobject.customizer.SqlStatementParameterCustomizer;
+package org.jdbi.v3.sqlobject.customizer.internal;
 
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Method;
 import java.lang.reflect.Parameter;
-import org.jdbi.v3.sqlobject.customizer.internal.BindFactory;
+import java.lang.reflect.Type;
+import org.jdbi.v3.sqlobject.customizer.Bind;
+import org.jdbi.v3.sqlobject.customizer.SqlStatementCustomizerFactory;
+import org.jdbi.v3.sqlobject.customizer.SqlStatementParameterCustomizer;
+import org.jdbi.v3.sqlobject.internal.ParameterUtil;
 
-/**
- * Standard implementation of {@link ParameterCustomizerFactory} that delegates to
- * {@link BindFactory#createForParameter(Annotation, Class, Method, Parameter, int, Type)} passing null for Annotation parameter.
- */
-public class BindParameterCustomizerFactory implements ParameterCustomizerFactory {
-
-    private SqlStatementCustomizerFactory bindSqlStatementCustomizerFactory = new BindFactory();
-
+public class BindFactory implements SqlStatementCustomizerFactory {
     @Override
-    public SqlStatementParameterCustomizer createForParameter(Class<?> sqlObjectType,
+    public SqlStatementParameterCustomizer createForParameter(Annotation annotation,
+                                                              Class<?> sqlObjectType,
                                                               Method method,
                                                               Parameter param,
                                                               int index,
                                                               Type type) {
-        return bindSqlStatementCustomizerFactory.createForParameter(null, sqlObjectType, method, param, index, type);
+        Bind b = (Bind) annotation;
+        String nameFromAnnotation = b == null ? Bind.NO_VALUE : b.value();
+        final String name = ParameterUtil.getParameterName(b, nameFromAnnotation, param);
+
+        return (stmt, arg) -> {
+            stmt.bindByType(index, arg, type);
+            stmt.bindByType(name, arg, type);
+        };
     }
 }
