+++ stringtemplate4/src/main/java/org/jdbi/v3/stringtemplate4/UseStringTemplateSqlLocator.java
@@ -13,27 +13,15 @@
  */
 package org.jdbi.v3.stringtemplate4;
 
-import static org.jdbi.v3.stringtemplate4.StringTemplateSqlLocator.findStringTemplateGroup;
-
-import java.lang.annotation.Annotation;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
-import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
 
-import org.jdbi.v3.core.config.ConfigRegistry;
-import org.jdbi.v3.core.statement.SqlStatements;
 import org.jdbi.v3.core.rewriter.ColonPrefixStatementRewriter;
 import org.jdbi.v3.core.rewriter.StatementRewriter;
-import org.jdbi.v3.sqlobject.internal.SqlAnnotations;
-import org.jdbi.v3.sqlobject.SqlObjects;
-import org.jdbi.v3.sqlobject.config.Configurer;
 import org.jdbi.v3.sqlobject.config.ConfiguringAnnotation;
-import org.jdbi.v3.sqlobject.locator.SqlLocator;
-import org.stringtemplate.v4.ST;
-import org.stringtemplate.v4.STGroup;
+import org.jdbi.v3.stringtemplate4.internal.UseStringTemplateSqlLocatorImpl;
 
 /**
  * Configures SQL Object to locate SQL using the {@link StringTemplateSqlLocator#findStringTemplate(Class, String)}
@@ -52,51 +40,9 @@ import org.stringtemplate.v4.STGroup;
  *     }
  * </pre>
  */
-@ConfiguringAnnotation(UseStringTemplateSqlLocator.Impl.class)
+@ConfiguringAnnotation(UseStringTemplateSqlLocatorImpl.class)
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ElementType.TYPE, ElementType.METHOD})
 public @interface UseStringTemplateSqlLocator {
     Class<? extends StatementRewriter> value() default ColonPrefixStatementRewriter.class;
-
-    class Impl implements Configurer {
-        @Override
-        public void configureForType(ConfigRegistry registry, Annotation annotation, Class<?> sqlObjectType) {
-            UseStringTemplateSqlLocator useStringTemplateSqlLocator = (UseStringTemplateSqlLocator) annotation;
-            SqlLocator locator = (type, method) -> {
-                String templateName = SqlAnnotations.getAnnotationValue(method).orElseGet(method::getName);
-                STGroup group = findStringTemplateGroup(type);
-                if (!group.isDefined(templateName)) {
-                    throw new IllegalStateException("No StringTemplate group " + templateName + " for class " + sqlObjectType);
-                }
-
-                return templateName;
-            };
-            StatementRewriter delegate = createDelegate(useStringTemplateSqlLocator.value());
-            StatementRewriter locatingRewriter = (sql, params, ctx) -> {
-                String templateName = sql;
-
-                STGroup group = findStringTemplateGroup(sqlObjectType);
-                ST template = group.getInstanceOf(templateName);
-                ctx.getAttributes().forEach(template::add);
-                String rewritten = template.render();
-
-                return delegate.rewrite(rewritten, params, ctx);
-            };
-            registry.get(SqlObjects.class).setSqlLocator(locator);
-            registry.get(SqlStatements.class).setStatementRewriter(locatingRewriter);
-        }
-
-        @Override
-        public void configureForMethod(ConfigRegistry registry, Annotation annotation, Class<?> sqlObjectType, Method method) {
-            configureForType(registry, annotation, sqlObjectType);
-        }
-
-        private StatementRewriter createDelegate(Class<? extends StatementRewriter> type) {
-            try {
-                return type.getConstructor().newInstance();
-            } catch (InstantiationException | IllegalAccessException | InvocationTargetException | NoSuchMethodException e) {
-                throw new IllegalStateException("Error instantiating delegate statement rewriter: " + type, e);
-            }
-        }
-    }
 }
