+++ stringtemplate4/src/main/java/org/jdbi/v3/stringtemplate4/UseStringTemplateStatementRewriter.java
@@ -13,53 +13,25 @@
  */
 package org.jdbi.v3.stringtemplate4;
 
-import java.lang.annotation.Annotation;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
-import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
 
-import org.jdbi.v3.core.config.ConfigRegistry;
-import org.jdbi.v3.core.statement.SqlStatements;
 import org.jdbi.v3.core.rewriter.ColonPrefixStatementRewriter;
 import org.jdbi.v3.core.rewriter.StatementRewriter;
-import org.jdbi.v3.sqlobject.config.Configurer;
 import org.jdbi.v3.sqlobject.config.ConfiguringAnnotation;
 import org.jdbi.v3.sqlobject.customizer.Define;
+import org.jdbi.v3.stringtemplate4.internal.UseStringTemplateStatementRewriterImpl;
 
 /**
  * Configures a SQL object class or method to rewrite SQL statements using StringTemplate. Method parameters annotated
  * with {@link Define @Define} are passed to the StringTemplate as template
  * attributes.
  */
-@ConfiguringAnnotation(UseStringTemplateStatementRewriter.Impl.class)
+@ConfiguringAnnotation(UseStringTemplateStatementRewriterImpl.class)
 @Target({ElementType.TYPE, ElementType.METHOD})
 @Retention(RetentionPolicy.RUNTIME)
 public @interface UseStringTemplateStatementRewriter {
     Class<? extends StatementRewriter> value() default ColonPrefixStatementRewriter.class;
-
-    class Impl implements Configurer {
-        @Override
-        public void configureForType(ConfigRegistry registry, Annotation annotation, Class<?> sqlObjectType) {
-            UseStringTemplateStatementRewriter useStringTemplateStatementRewriter = (UseStringTemplateStatementRewriter) annotation;
-            StatementRewriter delegate = createDelegate(useStringTemplateStatementRewriter.value());
-            StatementRewriter rewriter = new StringTemplateStatementRewriter(delegate);
-            registry.get(SqlStatements.class).setStatementRewriter(rewriter);
-        }
-
-        @Override
-        public void configureForMethod(ConfigRegistry registry, Annotation annotation, Class<?> sqlObjectType, Method method) {
-            configureForType(registry, annotation, sqlObjectType);
-        }
-
-        private StatementRewriter createDelegate(Class<? extends StatementRewriter> type) {
-            try {
-                return type.getConstructor().newInstance();
-            } catch (InstantiationException | IllegalAccessException | InvocationTargetException | NoSuchMethodException e) {
-                throw new IllegalStateException("Error instantiating delegate statement rewriter: " + type, e);
-            }
-        }
-    }
 }
